requirements.txt
asyncio
websockets
numpy
mysql-connector